+++ berserk/clients/bots.py
@@ -4,8 +4,8 @@ from typing import Iterator, Any, Dict
 
 from .. import models
 from ..formats import NDJSON
-from ..types.common import ChallengeDeclineReason
 from .base import BaseClient
+from ..types.challenges import ChallengeDeclineReason
 
 
 class Bots(BaseClient):